login_url correction with retention request detail page
Index views with permissionRequiredMixing had overridden
the login_url to an invalid location, causing error when
a user attempted to visit the page when not logged in.

Retention Request detail pages were also throwing an
error if a project had not been specified, even though
project is not a required field for the request.